+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Add interface definitions for HearthstoneJSON response
 
 ## [0.2.1] - 2016-12-14
 ### Fixed